URL params removed on browser navigation
