+++ package.json
@@ -12,7 +12,6 @@
     "cookie": "^0.4.0",
     "cross-fetch": "^3.0.3",
     "elasticsearch": "^16.6.0",
-    "history": "^5.0.0",
     "honeybadger-js": "^2.1.3",
     "jest-util": "25.1.0",
     "jquery": "^3.5.0",
